refactor: centralize v2session type
