+++ apps/app/types/logto-auth.d.ts
@@ -1,0 +1,11 @@
+interface V2Session {
+  logto_id: string;
+  provider: string;
+  providerId: string | number;
+  name: string;
+  image: string;
+  logto_username: string;
+  blacklisted: boolean;
+}
+
+export { V2Session };
